backend/multimodal/training/requirements_training.txt
# Core Deep Learning Frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Diffusers for Image Generation (FREE)
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0

# Computer Vision (FREE)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
scikit-image>=0.22.0

# Face Analysis (FREE - InsightFace)
insightface>=0.7.3
onnx>=1.15.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # For GPU acceleration

# Face Enhancement (FREE - GFPGAN)
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0

# Audio Processing (FREE)
TTS>=0.20.0  # Coqui TTS
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# 3D Generation (FREE)
trimesh>=4.0.0
pytorch3d>=0.7.5

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Progress & Logging
tqdm>=4.66.0
wandb>=0.16.0  # Weights & Biases for experiment tracking

# Utilities
requests>=2.31.0
huggingface-hub>=0.19.0
safetensors>=0.4.0
einops>=0.7.0

# Optional: Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Optional: Advanced Training
bitsandbytes>=0.41.0  # For 8-bit optimization
xformers>=0.0.23  # Memory efficient attention
